[CORE] Merge ConfigDefaults into config.hpp and add Doxygen docs
