+++ engine/core/CMakeLists.txt
@@ -20,7 +20,6 @@ add_library(EduGameEngineCore STATIC
     src/time.cpp
     src/config.cpp
     src/config_loader.cpp
-    src/config_default.cpp
 )
 
 # -------------------------
