+++ engine/core/src/config.cpp
@@ -1,7 +1,7 @@
-#include "engine/core/config.hpp"
+#include "engine/core/config/config.hpp"
 
 #include "engine/core/base.hpp"
-#include "engine/core/config_loader.hpp"
+#include "engine/core/config/config_loader.hpp"
 #include "engine/core/logging.hpp"
 #include "engine/core/types.hpp"
 
@@ -20,8 +20,22 @@ auto EngineConfig::saveToFile(const IConfigLoader& loader, const String& path) c
 
 void EngineConfig::intialize()
 {
+  setUpDefaultSettings();
   loadFromFile(JSONConfigLoader{}, String(ConfigDefaults::PATH_TO_CONFIG));
   EGE_INFO("Loading settings from JSON file completed...");
+}
+
+void EngineConfig::setUpDefaultSettings()
+{
+  const auto& defaultSettingsMap = ConfigDefaults::getDefaultSettingsMap();
+
+  for (const auto& [mapProperty, mapInfo] : defaultSettingsMap) {
+    std::visit(
+        [this, property = mapProperty](auto&& defaultValue) {
+          this->setProperty<std::decay_t<decltype(defaultValue)>>(property, std::forward<decltype(defaultValue)>(defaultValue));
+        },
+        mapInfo.defaultValue);
+  }
 }
