+++ engine/src/main.cpp
@@ -6,7 +6,7 @@
 #include <engine/ecs/world.hpp>
 #include <engine/platform/application_manager.hpp>
 
-#include "engine/core/config.hpp"
+#include "engine/core/config/config.hpp"
 #include "engine/core/types.hpp"
 
 using namespace engine;
