+++ settings.json
@@ -1,0 +1,1 @@
+{"PLATFORM":{"logLevel":"WARNING","targetFPS":120,"title":"NEW TITLE","windowHeight":720,"windowWidth":1234}}
